Pre-calculate partitions in partition_dups
Currently, we calculate the duplicate chain partitions in partition_dups_1.
This is done twice, once for phase 1 and once for phase 2.

Instead, calculate the partitions twice in partition_dups, once to calculate
the number of partitions (in order to allocate end/start/cnt arrays), and once
to initialize those arrays, and use those arrays in partition_dups_1 instead.

This allows us to reorder the duplicate chains after partitioning.

2020-01-21  Tom de Vries  <tdevries@suse.de>

	* dwz.c (partition_dups_1): Add and use end/start/cnt parameters.
	(count_partitions, initialize_partitions): New function.
	(partition_dups): Calculate partitions in arrays end/start/cnt, and
	pass these to partition_dups.