Also call mark_singletons for merged singleton die
When using --odr-mode=link, after splitting up an odr duplicate chain, if we
have an ODR_DECL chain and one or more ODR_DEF chains, we merge the ODR_DECL
chain with the first ODR_DEF chain.

If that ODR_DEF chain was a singleton chain, it's no longer called with
mark_singletons, while that is still necessary.

Fix this by detecting merged singleton DIEs and calling mark_singletons.

2020-01-21  Tom de Vries  <tdevries@suse.de>

	PR dwz/25424
	* dwz.c (merged_singleton): New function.
	(partition_dups): Call mark_singletons for merged_singleton die.