Add --devel-force
We can have the following message from dwz:
...
$ dwz a.out
dwz: a.out: DWARF compression not beneficial \
  - old size 69158054 new size 69158333
...

Given that the compression is not beneficial, no output is generated. However,
it may be interesting to have a look at the output, to understand why the
compression is not beneficial.

Add a developer-only option --devel-force that forces generation of output,
even if compression is not beneficial.

ChangeLog:

2020-02-13  Tom de Vries  <tdevries@suse.de>

	* dwz.c (force_p): New var.
	(dwz): Use force_p.
	(dwz_options, usage): Add --devel-force entry.