[odr] Guard odr code with odr_active_p
The odr optimization is enabled by the odr variable.

When enabled, some DIEs will get an die_odr_state of ODR_DEF or ODR_DECL,
which is then handled further by code guarded with odr.

However, it may be the case that all DIEs were marked with ODR_NONE, because:
- read_debug_info was run in low-mem mode, or
- read_debug_info was run in optimize-multifile mode, or
- there were no DIEs for which the optimization was applicable.

If all DIEs were marked with ODR_NONE, there is no point in entering further
odr handling code.

Add a variable odr_active_p that is set to true when at least one DIE has an
die_odr_state that is not ODR_NONE, and use the variable to guard some odr code.

2020-02-13  Tom de Vries  <tdevries@suse.de>

	* dwz.c (odr_active_p): New var.
	(set_die_odr_state): Set odr_active_p to true.
	(read_debug_info): Initialize odr_active_p to false.
	(partition_dups_1, partition_dups): Use odr_active_p instead of odr.