Add --devel-uni-lang
Add a developer-only option --devel-uni-lang, that forces duplicate chains to
consist of DIEs from CUs with the same language.

Note when using the option, each PU root DIE has the language attribute set.
This is done to make multifile work with this option (though the attribute is
also added in regular mode).  Alternatively, we could store that information
in a side table, to be kept inbetween multifile phases.

We will show here the effect of the option on the cc1 benchmark:
...
$ dwz -lnone cc1 -o 1 --devel-no-uni-lang
$ dwz -lnone cc1 -o 2 --devel-uni-lang
...

The benchmark consists of CUs with the following languages:
...
$ readelf -wi cc1 \
  | grep "DW_AT_language.*:" \
  | awk '{print $5, $6}' \
  | sort \
  | uniq -c
      2 (ANSI C)
     68 (ANSI C99)
    622 (C++)
      3 (MIPS assembler)
...

Doing the same analysis for 1 gives the same results, but for 2 we see the
additional language attributes in the PUs:
...
      2 (ANSI C)
    155 (ANSI C99)
   2753 (C++)
      3 (MIPS assembler)
...

The size effect for this example seems to be in the noise range:
...
$ diff.sh cc1 1
.debug_info      red: 44.84%    111527248  61527733
.debug_abbrev    red: 40.28%    1722726  1028968
.debug_str       red: 0%        6609355  6609355
total            red: 42.30%    119859329 69166056
$ diff.sh cc1 2
.debug_info      red: 44.84%    111527248  61521767
.debug_abbrev    red: 39.99%    1722726  1033916
.debug_str       red: 0%        6609355  6609355
total            red: 42.30%    119859329 69165038
...

Tested on-by-default with gdb testsuite using board cc-with-dwz and
cc-with-dwz-m.

2020-02-18  Tom de Vries  <tdevries@suse.de>

	* dwz.c (uni_lang_p): New var.
	(checksum_die, die_eq_1): Handle uni_lang_p.
	(read_debug_info): Ensure that the language attribute is read for
	uni_lang_p, and for PUs.
	(nr_bytes_for): New function.
	(partition_dups_1): Handle uni_lang_p in heuristics.
	(build_abbrevs_for_die): Add language attribute.
	(write_unit_die): Handle language attribute.
	(dwz_options, usage): Add --devel-uni-lang/--devel-no-uni-lang entries.