Add --devel-gen-cu
The DWARF standard appendix E.1 describes techniques that can be used for
compression and deduplication.

It explains that DIEs can be factored out into a PU (a compilation unit with
root DIE using DW_TAG_partial_unit), provided the PUs are then imported by
the CUs from which the DIEs where factored out.  This is what dwz implements.

But it explains as well that DIEs can be factored out into a CU (a compilation
unit with root DIE using DW_TAG_compile_unit), which dwz currently doesn't
implement.  This only works for DIEs that are globally visible, something
that is not generally valid, but is the case for C++.  The benefit of
factoring out into CUs is that imports are not required.

Given a partition of duplication chains, if the set of originating CUs are
all C++ CUs, then use a DW_TAG_compile_unit instead of a DW_TAG_partial_unit,
and don't use imports.

We will show here the effect of the option on the cc1 benchmark:
...
$ dwz -lnone cc1 -o 1 --devel-no-uni-lang --devel-no-gen-cu
$ dwz -lnone cc1 -o 2 --devel-no-uni-lang --devel-gen-cu
$ dwz -lnone cc1 -o 3 --devel-uni-lang --devel-no-gen-cu
$ dwz -lnone cc1 -o 4 --devel-uni-lang --devel-gen-cu
...

The benchmark consists of this many CUs:
...
$ readelf -wi cc1 \
  | egrep "\(DW_TAG_(compile|partial)_unit" \
  | awk '{print $5}' \
  | sort \
  | uniq -c
    695 (DW_TAG_compile_unit)
...

After normal compression, we have (file 1):
...
    695 (DW_TAG_compile_unit)
   4206 (DW_TAG_partial_unit)
...

Using --devel-gen-cu (file 2), we get the same result.

Using --devel-uni-lang (file 3), we get slightly less partial units:
...
    695 (DW_TAG_compile_unit)
   4143 (DW_TAG_partial_unit)
...

And when using both --devel-uni-lang and --devel-gen-cu (file 4), we get:
...
   3820 (DW_TAG_compile_unit)
    121 (DW_TAG_partial_unit)
...

The size effect for this example seems to be in the noise range:
...
$ diff.sh cc1 1
.debug_info      red: 44.84%    111527248  61527733
.debug_abbrev    red: 40.28%    1722726  1028968
.debug_str       red: 0%        6609355  6609355
total            red: 42.30%    119859329 69166056
$ diff.sh cc1 2
.debug_info      red: 44.84%    111527248  61527733
.debug_abbrev    red: 40.28%    1722726  1028968
.debug_str       red: 0%        6609355  6609355
total            red: 42.30%    119859329 69166056
$ diff.sh cc1 3
.debug_info      red: 44.84%    111527248  61521767
.debug_abbrev    red: 39.99%    1722726  1033916
.debug_str       red: 0%        6609355  6609355
total            red: 42.30%    119859329 69165038
$ diff.sh cc1 4
.debug_info      red: 44.89%    111527248  61463616
.debug_abbrev    red: 40.20%    1722726  1030280
.debug_str       red: 0%        6609355  6609355
total            red: 42.35%    119859329 69103251
...

Tested on-by-default, in combination with --devel-uni-lang on-by-default.

In that configuration, tested:
- regression testsuite, and
- gdb testsuite using board cc-with-dwz and cc-with-dwz-m.

With the external regression test suite, we see this regression:
...
Running src/testsuite/dwz-external.tests/dwz-external-tests.exp ...
dwz: 1: DWARF compression not beneficial - old size 69210922 new size 69428282
FAIL: src/testsuite/dwz-external.tests/pr24204.sh
...

The test-case uses cc1.dwz-processed, a cc1 binary that has already been
dwz-compressed using an older dwz.  I presume the FAIL to be due to
PR25566 - "[dwz] Take reference size into account in heuristics".

2020-02-19  Tom de Vries  <tdevries@suse.de>

	* dwz.c (gen_cu_p): New variable.
	(partition_lang): New function.
	(partition_dups_1): Create compilation unit with root DIE using tag
	DW_TAG_compile_unit, if allowed.  Update heuristics.
	(create_import_tree): Skip compilation unit with root DIE using tag
	DW_TAG_compile_unit.  Assert no imports are generated for such
	compilation units.
	(build_abbrevs_for_die): Add DW_AT_language attribute for compilation
	unit root DIE using tag DW_TAG_compile_unit.