Write DWARF5 multifile .debug_line if possible.
This write version 5 .debug_line into the multifile if all .debug_line
segments seen before are also version 5. This is more efficient than
version 2, if we don't have to write out the time and size of the files.
But only if there are more than ~25 files because the header is bigger
and there is some overhead for having to handle zero entry dir and file
entries (those do have to exist, but are not directly referenced by
anything).

ChangeLog:

	* dwz.c (lowest_line_version): New static unsigned int, set
	to 5 initially.
	(read_debug_line): Check version and set lowest_line_version.
	(struct line_stats): New.
	(list_line_entries): Update data as struct line_stats.
	(write_multifile_line): Collect time and size stats and write
	out version 5 .debug_line data if lowest_line_version is 5.