Break out of while loop correctly to make sure loclists are adjusted.
The reading and adjusting of DWARF5 .debug_loclists was modelled on
the DWARF4 .debug_loc section parsing. The .debug_loc parsing used
a while loop and breaks out when done. But .debug_loclists use a
switch statement inside the  while loop, so break doesn't actually
break out of the loop when done, and return on end of list exits
the functions too early. This meant that although the loclists were
parsed correctly, then were then not actually adjusted. Fix this
by using gotos inside the switch statements.

	* dwz.c (read_loclist_low_mem_phase1): Use again and done
	labels to goto inside switch.
	(read_loclist): Likewise.
	(adjust_loclist): Likewise.