Don't crash on DWARF5 .debug_line table with zero files.
Not having any files in a DWARF5 .debug_line table is odd, but not
technically invalid. Normally there should be a zero file entry that
is equal to the compile unit main file (there is no such zero entry
for early DWARF versions and we handle no file entries fine in that
case).

https://bugzilla.redhat.com/show_bug.cgi?id=1919243