Add DIE offsets in error messages to make it easier to find what is wrong.
With the following patch dwz will give a message like:

libmozjs-78.so: Couldn't find DIE at [bd6b507] referenced by
  DW_AT_abstract_origin from DIE at [bd5bb9b]

Which makes it a easier to figure out what is going on.

In the above case you can simply lookup the producer of the CU for those
two DIEs. Which turned out the be "clang LLVM (rustc version 1.49.0)"
which seems to have gotten the abstract origin reference wrong.

	* dwz.c (read_exprloc): Add DIE offsets to error messages.
	(checksum_die): Likewise.