Support DW_LANG_C_plus_plus_{03,11,14}
When compiling with f.i. g++ 7.5.0, we have GNU C++14 by default:
...
<d3> DW_AT_producer: GNU C++14 7.5.0 -mtune=generic -march=x86-64 -g
...
and the corresponding language setting is:
...
<d7> DW_AT_language: 4        (C++)
...

However, with -gdwarf-5, we have instead:
...
<d4> DW_AT_producer: GNU C++14 7.5.0 -mtune=generic -march=x86-64 -gdwarf-5
<d8> DW_AT_language: 33       (C++14)
...

The current checks in dwz (for --odr and --devel-gen-cu) for language C++
compare only with DW_LANG_C_plus_plus.

Fix this by comparing in addition with DW_LANG_C_plus_plus_{03,11,14}.

2021-02-14  Tom de Vries  <tdevries@suse.de>

	* dwz.c (set_die_odr_state, partition_lang): Add support for
	DW_LANG_C_plus_plus_{03,11,14}.