Add attribute to checksum for DW_FORM_implicit_const
In checksum_die, we handle DW_FORM_implicit_const like so:
...
       case DW_FORM_implicit_const:
         if (!handled && die->die_ck_state != CK_BAD)
           {
             handled = true;
             die->u.p1.die_hash
               = iterative_hash_object (t->values[i], die->u.p1.die_hash);
           }
...

The problem is that the attribute is not hashed in.

We could fix this by removing the "handled = true" assignment, but
generally the order is "attribute first, value second", and this fix would
do the reverse.  Consequently, identical attributes, one with form
DW_FORM_sdata and one with form DW_FORM_implicit_const would not get the same
checksum.

Fix this by locally adding hashing of the attribute, before hashing in the
value.

2021-02-15  Tom de Vries  <tdevries@suse.de>

	PR dwz/27418
	* dwz.c (checksum_die): Hash in attribute for DW_FORM_implicit_const.