Don't call die_odr_state with unnecessarily defined cu arg
When compiling dwz with this patch:
...
 die_odr_state (dw_cu_ref cu, dw_die_ref die)
 {
   if (die->die_odr_state != ODR_UNKNOWN)
-    return die->die_odr_state;
+    {
+      assert (cu == NULL);
+      return die->die_odr_state;
+    }
...
and running f.i. odr-struct.sh, we run into the abort.

The recent commit 3312feb "Fix CK_BAD propagation for --odr" introduced this
code:
...
                 if (die_odr_state (die_cu (die), die) != ODR_NONE)
                   die->u.p1.die_ref_hash = die->u.p1.die_hash;
...
and there's no need to pass a CU argument, which makes the abort trigger.

Fix this by passing a NULL CU instead.

2021-02-22  Tom de Vries  <tdevries@suse.de>

	* dwz.c (read_debug_info): Pass NULL CU to die_odr_state call.