Clean up temporary file in hardlink mode
Consider an executable file with hardlinks a.out and b.out.

When running dwz once:
...
$ dwz -h a.out b.out
...
a.out and b.out are updated, and remain hardlinks to the same file.

But when running dwz once more, a.out and b.out remain unchanged, and a
temporary file b.out.#dwz#.XXXXXX is left.

This is caused by the fact that the code in function dwz that is intended to
handle unchanged hardlinks is never triggered.  It is guarded by a
"resa[n].res == 1" condition, but res->res is set to 0 at the end of function
dwz, irrespective of whether the file changed or not.

Fix this by only setting res->res to 0 if the file changed.

This makes test-case twice-multifile.sh fail.  The test-case first
dwz-compresses files 1 and 2 in single-file mode.  Then it attempts a
dwz -m 3 1 2.

This results in a res.res value of 1 (processed, unchanged) for both files.

This causes the files not to be counted as a success in main, and multifile
mode optimization is skipped.

Fix this by handling res.res == 1 in main.

2021-03-02  Tom de Vries  <tdevries@suse.de>

	PR dwz/24275
	* dwz.c (dwz): Only set res->res to 0 if the file changed.
	(main): Handle res.res == 1.
	* testsuite/dwz.tests/twice-hardlink.sh: Remove PR24275 workaround.