Fix .debug_line reference above end of section
Consider the file test-import-repeatedly from PR26738, combined with a trivial
a.out:
...
$ cp test-import-repeatedly 1
$ gcc -m32 -g ~/hello.c
$ cp a.out 2
...

When doing multifile optimization, we run into:
...
$ dwz -m 3 2 1
dwz: 3: .debug_line reference above end of section
...

Using --devel-save-temps and src/contrib/gen-dwz-debug-all.sh we get the
unoptimized multifile, and find there a CU:
...
  Compilation Unit @ offset 0x371:
   Length:        0x6d (32-bit)
   Version:       4
   Abbrev Offset: 0x14d
   Pointer Size:  4
 <0><37c>: Abbrev Number: 1 (DW_TAG_compile_unit)
    <37d>   DW_AT_stmt_list   : 0xda
    <381>   DW_AT_language    : 0       (Unknown: 0)
    <382>   DW_AT_comp_dir    : ./build-3.8
...
which refers to a .debug_line offset 0xda, but the debug_line section only
contains an entry at offset 0x0.

This can be explained as follows.  The DIEs written into the CU do not
contain a single DW_AT_decl_file.  Consequently, htab_line will be NULL once
we get here in write_multifile:
...
             || (line_htab != NULL && write_multifile_line ()))
...
and no .debug_line contribution will be added.  However, the CU DIE
referencing the .debug_line contribution is written regardless.

We could fix this by not writing the DW_AT_stmt_list for the CU DIE in this
case, but that would require changing the order in which abbreviations are
computed.

Instead, fix this conservatively, by removing "line_htab != NULL &&", such we
get a minimal .debug_line contribution.

2021-03-02  Tom de Vries  <tdevries@suse.de>

	PR dwz/26738
	* dwz.c (write_multifile): Also call write_multifile_line when
	line_htab == NULL.