[testsuite] Fix odr-loc.sh with -gdwarf-5
When running the test-suite like this:
...
$ make clean; make; make check CC="gcc -gdwarf-5" CXX="g++ -gdwarf-5"
...
we run into:
...
FAIL: src/testsuite/dwz.tests/odr-loc.sh
...

The test-case calls dwz twice, once without and once with
--devel-ignore-locus.

The first call is there to verify that all DIEs that can be deduplicated
without --devel-ignore-locus are indeed deduplicated, but not more.

The FAIL happens because some DIEs are not deduplicated.  This is a regression
since commit 61c8d81 "Fix DW_AT_decl_file for odr".  This makes sense: the
commit changed partitioning, and now the size heuristics are preventing some
deduplication.

Fix this by ignoring the size heuristics in the first dwz call using
--devel-ignore-size.

2021-03-05  Tom de Vries  <tdevries@suse.de>

	* testsuite/dwz.tests/odr-loc.sh: Call dwz first time using
	--devel-ignore-size.