Add --devel-no-checksum-cycle-opt
In checksum_ref_die, we need to determine a per-toplevel-die checksum based on
the inter-toplevel-die references.  It has four modes:
- mode 1, a trivial one in which we walk towards toplevel-dies.
- mode 2, in which we optimistically assume that there are no cycles.
  If that is the case, we're done.  Otherwise, we use either mode 3 or mode 4.
- mode 3, dealing with cycles in a faster way
- mode 4, dealing with cycles

I did some testing using the testsuite and external testsuite and found that
mode 4 was not triggered once.  I'm not quite sure how to construct a
test-case for that either.

Add an option --devel-no-checksum-cycle-opt that can be used to enforce using
mode 3 instead of mode 4, and add a test-case that uses it.

2021-02-23  Tom de Vries  <tdevries@suse.de>

	PR dwz/27429
	* Makefile (TEST_EXECS): Add cycle.
	(cycle): New target.
	* dwz.c (checksum_cycle_opt): New var.
	(checksum_ref_die): Handle checksum_cycle_opt.
	(dwz_options, usage): Add --devel-checksum-cycle-opt entries.
	* testsuite/dwz.tests/cycle.c: New test.
	* testsuite/dwz.tests/cycle.sh: New test.