[testsuite] Handle readelf following links by default
Since binutils commit c46b706620e "Change the readelf and objdump programs so
that they will automatically follow links to separate debug info files" there
are a few FAILs in the dwz testsuite.

This is just due to changing the default behaviour for readelf.

Fix this by testing whether the new -Wn readelf flag is supported,
and if so, adding it to the readelf call.

2021-03-20  Tom de Vries  <tdevries@suse.de>

	PR dwz/27592
	* testsuite/dwz.tests/odr-struct-multifile.sh: Call readelf with -wN
	if supported.
	* testsuite/dwz.tests/cycle.sh: Same.