Make smaller-than.sh match dwz heuristic
As reported in PR27603, when using clang-11 we run into:
...
Running testsuite/dwz.tests/dwz-tests.exp ...
+ exec=start-gold
+ cp start-gold 1
+ dwz 1
+ smaller-than.sh 1 start-gold
FAIL: testsuite/dwz.tests/pr24747.sh
...

The files are the same size:
...
$ ls -l start-gold tmp.pr24747.sh/1
-rwxr-xr-x 1 vries users 2024 Mar 18 11:58 start-gold
-rwxr-xr-x 1 vries users 2024 Mar 18 11:58 tmp.pr24747.sh/1
...
but not the same:
...
$ diff -q start-gold tmp.pr24747.sh/1
Files start-gold and tmp.pr24747.sh/1 differ
...

The .debug_info size is smaller, by 6 bytes:
...
 [Nr] Name         Type      Address  Off    Size   ES Flg Lk Inf Al
-[ 6] .debug_info  PROGBITS  00000000 0001cd 000040 00      0   0  1
+[ 6] .debug_info  PROGBITS  00000000 0001cd 00003a 00      0   0  1
...

The size reduction in the .debug_info section happens not to result in a
smaller file because subsequent sections .note.gnu.gold-version and .symtab
have a bigger-than-one align.

We could try to fix this by adapting the heuristic to not compress if the
eventual file size is equal.

Instead, fix this by making the smaller-than.sh test match the heuristic,
which only looks at the size of some .debug sections.

2021-03-22  Tom de Vries  <tdevries@suse.de>

	* testsuite/scripts/smaller-than.sh: Calculate size as sum of
	.debug sections.