Run two-files-low-mem-die-limit-0.sh with -j1
On the buildbot we have for dwz-debian-arm64:
...
child process exited abnormally
FAIL: build/testsuite/dwz.tests/two-files-low-mem-die-limit-0.sh
...

AFAIU, the problem is that the test-case is grepping like this:
...
egrep -q "Compressing (1|2)$" dwz.err
...
expecting the grep to fail on these messages:
...
Compressing 1 in low-mem mode
Compressing 2 in low-mem mode
...

But with -j 2 we can have the following interleaving of the messages:
...
Compressing 1 in low-mem modeCompressing 2
 in low-mem mode
...
which makes the grep succeed.

Fix this by running the test with -j1.

2021-03-26  Tom de Vries  <tdevries@suse.de>

	* testsuite/dwz.tests/two-files-low-mem-die-limit-0.sh: Run dwz
	with -j1.