Calculate workset before serial/parallel code in dwz_files_1
The function dwz_files_1 has code that can execute either serially or in
parallel.  Evidently, there's code duplication between the serial and
parallel code, which creates the burden of keeping this in sync.  The
burden can be kept minimal by keeping the duplicate code minimal.

Part of the duplicate code is the selection which files to handle.

Move this selection out of the serial/parallel code, and use it to compute
an array workset, which is then used in both the serial and parallel code.

2021-03-27  Tom de Vries  <tdevries@suse.de>

	* dwz.c (dwz_files_1): Calculate workset before serial/parallel code.