Fix incorrect res arg for wait_child_exit
In dwz_files_1 we have:
...
         file = files[i];
         struct file_result *res = &resa[i];

         if (nr_forks == max_forks)
           {
             int thisret = wait_child_exit (-1, pids, i, res);
...

The -1 pid argument to wait_child_exit means that any child status will be
returned, while the res argument passes the result array element for
the file we want to start processing, which is incorrect.

Fix this by passing the result array instead, and letting wait_child_exit
figure out which element to use.

2021-03-29  Tom de Vries  <tdevries@suse.de>

	* dwz.c (wait_child_exit): Find appropriate element in result array.
	(dwz_files_1): Use resa when using pid -1.