Enable parallel multifile with -e -p
Currently, parallel dwz is disabled when multifile is used:
...
$ dwz -m 5 3 1 2 4 -j 4
...

Enable this when the multifile parameter characteristics are specified using
-p and -e:
...
$ dwz -m 5 3 1 2 4 -j 4 -p 8 -e l
...
This works around the child processes having to communicate back to the parent
the found pointer size and endiannes, and doing the -j auto and -e auto
consistency checking.

The problem of the different child processes writing to the same temporary
multifile is solved by writing in file order to the temporary multifile.

In principle, the problem could be solved by writing into per-file temporary
multifiles and concatenating them afterwards.  However, the temporary
multifile contains DW_FORM_ref_addr references, and after concatenation
those references in the temporary multifile require relocation (i.e., add the
offset of the start of the file contribution).  This requires a lot of
changes in various parts of the code, so for now we choose instead this
cleaner solution.

The enforcing of writing in file order is done by passing a token to each
child process using pipes.

Experiment:
...
$ for j in 1 2; do \
    cp debug/cc1 1; cp 1 2; cp 2 3; cp 3 4; \
    echo "j: $j"; \
    ../measure/time.sh ./dwz -lnone -m 5 1 2 3 4 -j $j -p 8 -e l; \
  done
j: 1
maxmem: 1297260
real: 48.35
user: 46.93
system: 1.37
j: 2
maxmem: 1296584
real: 31.85
user: 50.46
system: 1.64
...

2021-03-30  Tom de Vries  <tdevries@suse.de>

	PR dwz/25951
	* dwz.c (write_multifile_1): Factor out of ...
	(write_multifile): ... here.  Call get_token.
	(get_token, pass_token): New function.
	(wait_children_exit): Call pass_token.
	(dwz_files_1): Allow parallel multifile with -j -e.  Call pass_token.