Eliminate sed usage in building dwz-for-test
We started out building dwz-for-test using -U__GNUC__.  That resulted
in problems with system headers on fedora, so instead a solution using
sed was implemented.  This works well and ensures that all __GNUC__ uses
are addressed, but it makes refactoring more complicated.

Instead, use the slightly more error-prone but less cumbersome method
of replacing uses of __GNUC__ with uses of a new variable USE_GNUC,
and define USE_GNUC=0 for dwz-for-test.

2021-04-09  Tom de Vries  <tdevries@suse.de>

	* Makefile: Compile dwz-for-test using -DUSE_GNUC=0.
	* dwz.c (USE_GNUC): Weakly define if __GNUC__ is defined, and
	use instead of __GNUC__.